fix: custom tool serializer
