+++ api/api/routers/mcp/_mcp_utils.py
@@ -1,0 +1,36 @@
+import re
+
+from api.routers.mcp._mcp_errors import MCPError
+
+
+def extract_agent_id_and_run_id(run_url: str) -> tuple[str, str]:
+    """Extract the agent ID and run ID from the run URL.
+
+    Supports multiple URL formats:
+    1. https://workflowai.com/workflowai/agents/classify-email-domain/runs/019763ae-ba9f-70a9-8d44-5a626c82e888
+    2. http://localhost:3000/workflowai/agents/sentiment/2/runs?taskRunId=019763a5-12a7-73b7-9b0c-e6413d2da52f
+
+    Args:
+        run_url: The run URL to parse
+
+    Returns:
+        A tuple of (agent_id, run_id)
+
+    Raises:
+        ValueError: If the URL format is invalid or doesn't match the expected pattern
+    """
+    if not run_url:
+        raise ValueError("run_url must be a non-empty string")
+
+    regexps = [
+        r"^.*/agents/([^/]+)/runs/([-_a-zA-Z0-9]+)",
+        r"^.*/agents/([^/]+)/\d+/runs?.*runId=([-_a-zA-Z0-9]+).*$",
+    ]
+
+    for regexp in regexps:
+        match = re.match(regexp, run_url, re.IGNORECASE)
+        if match:
+            return match.group(1), match.group(2)
+    raise MCPError(
+        "Invalid run URL, must be in the format 'https://workflowai.com/workflowai/agents/agent-id/runs/run-id', or you must pass 'agent_id' and 'run_id'",
+    )
