+++ api/api/routers/mcp/mcp_server.py
@@ -24,13 +24,14 @@ from api.routers.mcp._mcp_models import (
     SearchResponse,
     SortOrder,
 )
+from api.routers.mcp._mcp_serializer import tool_serializer
 from api.services.tools_service import ToolsService
 from core.domain.tool import Tool
 from core.domain.users import UserIdentifier
 from core.storage.backend_storage import BackendStorage
 from core.utils.schema_formatter import format_schema_as_yaml_description
 
-_mcp = FastMCP("WorkflowAI 🚀", stateless_http=True)  # pyright: ignore [reportUnknownVariableType]
+_mcp = FastMCP("WorkflowAI 🚀", tool_serializer=tool_serializer)  # pyright: ignore [reportUnknownVariableType]
 
 
 async def get_task_tuple_from_task_id(storage: BackendStorage, agent_id: str) -> TaskTuple:
@@ -206,10 +207,8 @@ async def fetch_run_details(
     - agent_schema_id: The schema/version ID of the agent used for this run
     - status: Current status of the run (e.g., "completed", "failed", "running")
     - conversation_id: Links this run to a broader conversation context if applicable
-
-    **Input/Output Data:**
-    - agent_input: Complete input data provided to the agent for this run
-    - agent_output: Complete output/response generated by the agent
+    - agent_input: Input data if any that was provided to the agent (only when using input variables)
+    - messages: The exchanged messages, including the returned assistant message
 
     **Performance Metrics:**
     - duration_seconds: Execution time in seconds
@@ -734,4 +733,4 @@ async def search_documentation(
 
 
 def mcp_http_app():
-    return _mcp.http_app(path="/")
+    return _mcp.http_app(path="/", stateless_http=True)
